+++ Soggins/Cards/Apologies/RedrawApologyCard.cs
@@ -1,0 +1,30 @@
+﻿using CobaltCoreModding.Definitions.ExternalItems;
+using CobaltCoreModding.Definitions.ModContactPoints;
+using System.Collections.Generic;
+
+namespace Shockah.Soggins;
+
+public sealed class RedrawApologyCard : ApologyCard, IRegisterableCard
+{
+	public void RegisterCard(ICardRegistry registry)
+	{
+		ExternalCard card = new(
+			globalName: $"{GetType().Namespace}.Card.Apology.Redraw",
+			cardType: GetType(),
+			cardArt: Art,
+			actualDeck: ModEntry.Instance.SogginsDeck
+		);
+		card.AddLocalisation(I18n.ApologyCardName);
+		registry.RegisterCard(card);
+	}
+
+	public override List<CardAction> GetActions(State s, Combat c)
+		=> [
+			new AStatus
+			{
+				status = ModEntry.Instance.KokoroApi.RedrawVanillaStatus,
+				statusAmount = 1,
+				targetPlayer = true
+			}
+		];
+}