+++ Soggins/ExternalAPIs/Kokoro/RedrawStatus.cs
@@ -1,0 +1,6 @@
+﻿namespace Shockah.Soggins;
+
+public partial interface IKokoroApi
+{
+	Status RedrawVanillaStatus { get; }
+}