+++ Soggins/docs/release-notes.md
@@ -2,6 +2,10 @@
 
 # Release notes
 
+## Upcoming release
+
+* Added a Redraw apology card.
+
 ## 1.3.3
 Released 5 May 2024.